Move stderr into its own module
